split into files, delete comment in parser
