+++ src/node.cpp
@@ -1,8 +1,6 @@
 #include "node.hpp"
 
 #include <algorithm>
-#include <map>
-#include <set>
 
 Node::Node(char op) : op_(op)
 {
@@ -108,6 +106,7 @@ string Node::getString()
         tmp_str += ")";
         return tmp_str;
     }
+    return "";
 }
 
 
@@ -156,160 +155,6 @@ Node *Node::simplifyTree()
     }
     if (children_.size() == 1)
     {
-        Node *new_root = children_[0];
-        if (parent_)
-        {
-            parent_->replaceChild(this, new_root);
-        }
-        new_root->setParent(parent_);
-        children_.clear();
-        delete this;
-        return new_root;
-    }
-    return this;
-}
-
-bool Node::checkEmptiness()
-{
-    if (op_ == ops::LETTER)
-    {
-        return false;
-    }
-    if (op_ == ops::STAR)
-    {
-        return true;
-    }
-    if (op_ == ops::CONCAT)
-    {
-        for (auto child : children_)
-        {
-            if (!child->checkEmptiness())
-            {
-                return false;
-            }
-        }
-        return true;
-    }
-    if (op_ == ops::OR)
-    {
-        for (auto child : children_)
-        {
-            if (child->checkEmptiness())
-            {
-                return true;
-            }
-        }
-        return false;
-    }
-    return false;
-}
-
-Node *Node::aci()
-{
-    if (op_ == ops::LETTER) {
-        return this;
-    }
-    if (op_ == ops::STAR) {
-        children_[0] = children_[0]->aci();
-        return this;
-    }
-    if (op_ == ops::CONCAT) {
-        for (auto ch : children_) {
-            ch = ch->aci();
-        }
-        return this;
-    }
-    if (op_ == ops::OR) {
-        map<string, Node*> mp_str_to_node;
-        multiset<Node*> extra_nodes;
-        for (auto ch : children_) {
-            ch = ch->aci();
-            mp_str_to_node[ch->getString()] = ch;
-            extra_nodes.insert(ch);
-        }
-        children_.clear();
-        for (auto pair : mp_str_to_node) {
-            children_.push_back(pair.second);
-            auto it = extra_nodes.find(pair.second);
-            extra_nodes.erase(it);
-        }
-
-        // удаление из динамической памяти лишние аргументы альтернативы
-        for (auto node : extra_nodes) {
-            node->clearChildren();
-            delete node;
-        }
-        return this;
-    }
-    return this;
-}
-
-Node *Node::ssnf()
-{
-    if (op_ == ops::LETTER)
-    {
-        return this;
-    }
-    if (op_ == ops::CONCAT || op_ == ops::OR)
-    {
-        for (auto ch : children_)
-        {
-            ch = ch->ssnf();
-        }
-        return this;
-    }
-    if (op_ == ops::STAR)
-    {
-        children_[0] = children_[0]->ss();
-        return this;
-    }
-    return this;
-}
-
-Node *Node::ss()
-{
-    if (op_ == ops::LETTER)
-    {
-        return this;
-    }
-    if (op_ == ops::OR)
-    {
-        for (auto ch : children_)
-        {
-            ch = ch->ss();
-        }
-        return this;
-    }
-    if (op_ == ops::CONCAT)
-    {
-        if (checkEmptiness())
-        {
-            Node *new_or = new Node(ops::OR);
-            for (auto ch : children_)
-            {
-                ch = ch->ss();
-                new_or->addChild(ch);
-                ch->setParent(new_or);
-            }
-            if (parent_)
-            {
-                parent_->replaceChild(this, new_or);
-            }
-            new_or->setParent(parent_);
-            children_.clear();
-            delete this;
-            return new_or;
-        }
-        for (auto ch : children_)
-        {
-            ch = ch->ssnf();
-        }
-
-        return this;
-    }
-    if (op_ == ops::STAR)
-    {
-        children_[0]->ss();
         Node *new_root = children_[0];
         if (parent_)
         {
